requirements.txt
# requirements.txt
pandas==2.1.3
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas-market-calendars==4.3.2
pytz==2024.1